+++ src/SleetLib/def/ISleetService.cs
@@ -3,17 +3,11 @@ using System.Threading.Tasks;
 
 namespace Sleet
 {
-    public interface ISleetService : IAddRemovePackages
+    public interface ISleetService : IApplyOperations
     {
         /// <summary>
         /// Service name
         /// </summary>
         string Name { get; }
-
-        /// <summary>
-        /// Fetch the needed service files. This can be used to load
-        /// all files in parallel.
-        /// </summary>
-        Task FetchAsync();
     }
 }