+++ test/SleetLib.Tests/CatalogTests.cs
@@ -11,9 +11,10 @@ using NuGet.Packaging;
 using NuGet.Packaging.Core;
 using NuGet.Test.Helpers;
 using NuGet.Versioning;
+using Sleet;
 using Xunit;
 
-namespace Sleet.Test
+namespace SleetLib.Tests
 {
     public class CatalogTests
     {
@@ -98,15 +99,11 @@ namespace Sleet.Test
                 var zipFile = testPackage.Save(packagesFolder.Root);
                 using (var zip = new ZipArchive(File.OpenRead(zipFile.FullName), ZipArchiveMode.Read, false))
                 {
-                    var input = new PackageInput(zipFile.FullName, new PackageIdentity("packageA", NuGetVersion.Parse("1.0.0-alpha.1")), false)
-                    {
-                        NupkgUri = UriUtility.CreateUri("http://tempuri.org/flatcontainer/packageA/1.0.0-alpha.1/packageA.1.0.0-alpha.1.nupkg"),
-                        Zip = zip,
-                        Package = new PackageArchiveReader(zip)
-                    };
-
+                    var input = PackageInput.Create(zipFile.FullName);
+                    var nupkgUri = UriUtility.CreateUri("http://tempuri.org/flatcontainer/packageA/1.0.0-alpha.1/packageA.1.0.0-alpha.1.nupkg");
+                    
                     // Act
-                    var actual = await CatalogUtility.CreatePackageDetailsAsync(input, catalog.CatalogBaseURI, context.CommitId, writeFileList: true);
+                    var actual = await CatalogUtility.CreatePackageDetailsAsync(input, catalog.CatalogBaseURI, nupkgUri, context.CommitId, writeFileList: true);
 
                     var dependencyGroups = actual["dependencyGroups"] as JArray;
                     var frameworkAssemblyGroups = actual["frameworkAssemblyGroup"] as JArray;
@@ -183,15 +180,11 @@ namespace Sleet.Test
                 var zipFile = testPackage.Save(packagesFolder.Root);
                 using (var zip = new ZipArchive(File.OpenRead(zipFile.FullName), ZipArchiveMode.Read, false))
                 {
-                    var input = new PackageInput(zipFile.FullName, new PackageIdentity("packageA", NuGetVersion.Parse("1.0.0")), false)
-                    {
-                        NupkgUri = UriUtility.CreateUri("http://tempuri.org/flatcontainer/packageA/1.0.0/packageA.1.0.0.nupkg"),
-                        Zip = zip,
-                        Package = new PackageArchiveReader(zip)
-                    };
-
+                    var input = PackageInput.Create(zipFile.FullName);
+                    var nupkgUri = UriUtility.CreateUri("http://tempuri.org/flatcontainer/packageA/1.0.0/packageA.1.0.0.nupkg");
+                    
                     // Act
-                    var actual = await CatalogUtility.CreatePackageDetailsAsync(input, catalog.CatalogBaseURI, context.CommitId, writeFileList: true);
+                    var actual = await CatalogUtility.CreatePackageDetailsAsync(input, catalog.CatalogBaseURI, nupkgUri, context.CommitId, writeFileList: true);
 
                     var dependencyGroups = actual["dependencyGroups"] as JArray;
                     var frameworkAssemblyGroups = actual["frameworkAssemblyGroup"] as JArray;
@@ -245,7 +238,7 @@ namespace Sleet.Test
                 };
 
                 var catalog = new Catalog(context);
-                var catalogIndex = TemplateUtility.LoadTemplate(
+                var catalogIndex = await TemplateUtility.LoadTemplate(
                     "CatalogIndex",
                     DateTimeOffset.UtcNow,
                     fileSystem.BaseURI);
@@ -262,20 +255,9 @@ namespace Sleet.Test
                 using (var zipA = new ZipArchive(File.OpenRead(zipFileA.FullName), ZipArchiveMode.Read, false))
                 using (var zipB = new ZipArchive(File.OpenRead(zipFileB.FullName), ZipArchiveMode.Read, false))
                 {
-                    var inputA = new PackageInput(zipFileA.FullName, new PackageIdentity("packageA", NuGetVersion.Parse("1.0.0")), false)
-                    {
-                        NupkgUri = UriUtility.CreateUri("http://tempuri.org/flatcontainer/packageA/1.0.0/packageA.1.0.0.nupkg"),
-                        Zip = zipA,
-                        Package = new PackageArchiveReader(zipA)
-                    };
-
-                    var inputB = new PackageInput(zipFileB.FullName, new PackageIdentity("packageB", NuGetVersion.Parse("1.0.0")), false)
-                    {
-                        NupkgUri = UriUtility.CreateUri("http://tempuri.org/flatcontainer/packageB/1.0.0/packageB.1.0.0.nupkg"),
-                        Zip = zipB,
-                        Package = new PackageArchiveReader(zipB)
-                    };
-
+                    var inputA = PackageInput.Create(zipFileA.FullName);
+                    var inputB = PackageInput.Create(zipFileB.FullName);
+                    
                     // Act
                     await catalog.AddPackageAsync(inputA);
                     await catalog.AddPackageAsync(inputB);